backend/requirements.txt
coverage==4.5.3
Django==2.1.7
djangorestframework==3.9.2
Pillow==5.4.1
pre_commit==1.14.4
pytest==4.3.0
pytest-cov==2.6.1
pytest-custom-exit-code==0.1.1
pytest-django==3.4.8
pytz==2018.9
tox==3.7.0